+++ src/evaldet/__init__.py
@@ -1,6 +1,6 @@
 """Evaluation for Detection and Tracking."""
 
-__version__ = "0.4.2"
+__version__ = "0.4.3"
 
 __all__ = ["Detections", "Tracks"]